fourth 4.3 commit - fixes of relations of entities and sql's querys - some docs of RestaurantController, RestaurantService and RestaurantRepository

+++ src/main/java/com/example/restaurantapi/service/RestaurantService.java
@@ -5,6 +5,7 @@ import com.example.restaurantapi.model.dto.RestaurantReqDto;
 import com.example.restaurantapi.model.entity.RestaurantModel;
 import com.example.restaurantapi.repository.RestaurantRepository;
 import com.example.restaurantapi.utils.exception.NoRestaurantFoundException;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,6 +17,14 @@ public class RestaurantService {
 
     private final RestaurantRepository restaurantRepository;
 
+    /**
+     *
+     * Creates new RestaurantModel object with the provided details.
+     *
+     * @param req The restaurant details as dto
+     * @return A new RestaurantModel object with provided details
+     */
+
     public RestaurantModel createRestaurant(NewRestaurantReq req) {
         RestaurantModel newRestaurant = new RestaurantModel();
         newRestaurant.setName(req.getName());
@@ -26,9 +35,19 @@ public class RestaurantService {
         return restaurantRepository.save(newRestaurant);
     }
 
+    /**
+     *
+     * Updates an existing restaurant with new details
+     *
+     * @param id the unique identifier of the restaurant to update
+     * @param req The new restaurant details as dto
+     * @return The update RestaurantModel object
+     * @throws NoRestaurantFoundException If no restaurant is found with the provided id
+     */
+
     public RestaurantModel updateRestaurant(Long id, RestaurantReqDto req) throws NoRestaurantFoundException {
         RestaurantModel restaurant = restaurantRepository.findById(id)
-                .orElseThrow( () -> new NoRestaurantFoundException("Not found nein"));
+                .orElseThrow( () -> new NoRestaurantFoundException("Can't update restaurant data because there is no restaurant with such an ID"));
         restaurant.setName(req.getName());
         restaurant.setType(req.getType());
         restaurant.setOpenHour(req.getOpenHour());
@@ -36,19 +55,55 @@ public class RestaurantService {
         return restaurantRepository.save(restaurant);
     }
 
+    /**
+     *
+     * Retrives a restaurant by its unique identifier
+     *
+     * @param restaurantName
+     * @return
+     * @throws NoRestaurantFoundException
+     */
     public RestaurantModel findByName(String restaurantName) throws NoRestaurantFoundException {
         Optional<RestaurantModel> optRestaurant = restaurantRepository.findByNameIsIgnoreCase(restaurantName);
 
         if(optRestaurant.isPresent()) {
-            RestaurantModel restuarant = optRestaurant.get();
-            return restuarant;
+            return optRestaurant.get();
         }
-        throw new NoRestaurantFoundException("Not found Restaurant , Hello from findByName");
+        throw new NoRestaurantFoundException("Can't find restaurant's data because there is no restaurant with such an ID");
     }
 
-    public void deleteRestaurantByName(Long id)
-    {
-        restaurantRepository.deleteById(id);
+    /**
+     *
+     * Retrives a restaurant by its unique identifier
+     *
+     * @param id
+     * @return
+     * @throws NoRestaurantFoundException If no restaurant is found with the provided id
+     */
+
+    public RestaurantModel findByAdressId(Long id) throws NoRestaurantFoundException {
+        Optional<RestaurantModel> optRestaurant = restaurantRepository.findByAddress_Id(id);
+        if( optRestaurant.isPresent()) {
+            return optRestaurant.get();
+        }
+        throw new NoRestaurantFoundException("Can't find restaurant's address because there is no address with such an ID");
+    }
+
+    /**
+     *
+     * Deletes a restaurant by its unique identifier
+     *
+     * @param id The unique
+     *
+     * @throws NoRestaurantFoundException if no restaurant is found with the provided id
+     */
+    public void deleteRestaurantById(Long id) throws NoRestaurantFoundException {
+        Optional<RestaurantModel> optRestaurant = restaurantRepository.findById(id);
+        if(optRestaurant.isPresent()) {
+            restaurantRepository.deleteById(id);
+        } else {
+            throw new NoRestaurantFoundException("Can't delete restaurant object because there is no restaurant with such an ID");
+        }
     }
 
 }
